Implemented fileinfo extension.
This extension uses the stock libmagic code, which is unfortunately
not thread-safe. This means that each fileinfo operation uses a global
lock, which is not ideal, but it is better than having no functionality
at all. I assume for most use cases, support for heavy parallelized
fileinfo usage will not be required.

Most of the phpt tests in version 5.3.8 have been ported to the C++
test suite.

As differences from php version, there is no support for streams and
there are more warning messages explaining why something failed instead
of the simple return false.